backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.8.0
PyMuPDF==1.23.8
werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
SpeechRecognition==3.10.0
soundfile==0.12.1
sounddevice==0.4.6
scipy==1.11.4
numpy==1.24.3
pydub==0.25.1
librosa==0.10.1
vosk==0.3.45
transformers==4.35.0
torch==2.0.1